requirements.txt
appdirs==1.4.3
CacheControl==0.12.5
certifi==2018.8.24
chardet==3.0.4
Click==7.0
colorama==0.3.9
dash==0.28.2
dash-core-components==0.33.0
dash-html-components==0.13.2
dash-renderer==0.14.3
decorator==4.3.0
distlib==0.2.8
distro==1.3.0
docutils==0.14
Flask==1.0.2
Flask-Compress==1.4.0
html5lib==1.0.1
idna==2.7
ipython-genutils==0.2.0
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
jupyter-core==4.4.0
lockfile==0.12.2
MarkupSafe==1.0
msgpack==0.5.6
nbformat==4.4.0
packaging==18.0
pipenv==2018.10.13
plotly==3.3.0
progress==1.4
pyparsing==2.2.2
pytoml==0.1.19
pytz==2018.5
requests==2.19.1
retrying==1.3.3
six==1.11.0
smbus2==0.2.1
traitlets==4.3.2
urllib3==1.23
virtualenv==16.0.0
virtualenv-clone==0.3.0
webencodings==0.5.1
Werkzeug==0.14.1
gunicorn==19.9.0
psycopg2-binary==2.7.5
